Separate OrganizeImports into its own namespace and file
